UC4.6 - changed permissions field from user to storeManager and storeOwner
