+++ workshopCode/tests/AcceptanceTests/UC4_10.java
@@ -1,0 +1,5 @@
+package AcceptanceTests;
+
+public class UC4_10 {
+
+}
